+++ 3x3x3/fonctions.h
@@ -6,4 +6,4 @@ void melangeCube(int cube[6][3][3]);
 void copie(int cubeACopier[6][3][3], int cubeAColler[6][3][3]);
 
 //Fonctions du temps
-double stopClock(clock_t tempsDebut);
+float stopClock(clock_t tempsDebut);
